+++ samples/net/vrbox/src/service.h
@@ -24,7 +24,7 @@ int service_cmd_parse(uint8_t *msg, size_t mes_len);
 int service_send_error_log(const char *msg);
 
 #ifdef CONFIG_APP_FACTORY_TEST
-bool service_cmd_is_factory_test(uint8_t*, size_t);
+bool service_cmd_is_factory_test(uint8_t*, uint16_t);
 #endif /* CONFIG_APP_FACTORY_TEST */
 
 #ifdef __cplusplus
